backend/requirements.txt
# FastAPI and Server
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
python-dotenv==1.0.1

# CORS and Middleware
fastapi-cors==0.0.6

# Data Processing and ML
numpy
pandas
scikit-learn
joblib==1.4.2

# Database Drivers
redis==5.2.0
influxdb-client==1.47.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
databases[postgresql]==0.9.0  # Async database support for FastAPI

# MQTT
paho-mqtt==2.1.0

# Blockchain
web3==7.6.0
eth-account==0.13.4


# HTTP Requests
aiohttp==3.11.10
requests==2.32.3
httpx==0.27.0

# Configuration Management
pydantic==2.10.3
pydantic-core==2.27.1
pydantic-settings==2.6.1

# Async Support
asyncio==3.4.3


# Deep Learning & NLP
torch
torchvision
transformers
pillow==11.3.0
pillow==11.3.0

# Logging and Monitoring
python-json-logger==3.2.1
# Trigger build
